+++ src/main/java/com/bing/lan/invest/domain/entity/AccountFundDetail.java
@@ -2,8 +2,10 @@ package com.bing.lan.invest.domain.entity;
 
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.bing.lan.invest.base.BaseEntity;
+
 import java.io.Serializable;
 import java.math.BigDecimal;
+
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
@@ -63,7 +65,7 @@ public class AccountFundDetail extends BaseEntity {
     private BigDecimal custPart;
 
     @ApiModelProperty("计划份数")
-    private Long planPart;
+    private BigDecimal planPart;
 
     @ApiModelProperty("持有份额")
     private BigDecimal totalShare;
